requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
pypdf
streamlit
python-dotenv